requirements.txt
python-telegram-bot>=21.6
requests
flask
gTTS
google-generativeai